code/src/requirements.txt
# Core Packages
streamlit==1.28.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.0
Pillow==10.0.1

# AI/ML Models
transformers==4.34.0
torch==2.1.0
torchvision==0.16.0

# Data Visualization
matplotlib==3.8.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
tempfile==1.0.0
uuid==1.30